feat: updated layout to handle mobile better
